Implementation of whole thing
Sorry for not dividing it into commits. Philosopher databse test re-added, since the implementation of non recursive rules and recursive rules in my implementation are the same. I did not use records because i saw no need of them, i dislike the idea of having the database in memory. Changing to file database instead of single string should be simple, marked lines where changes are neccesary to do so. It is not a very performant solution, but an index to not perform a linear search could fix that.